+++ programs-ecs/bundles/ecs-bundle/src/lib.rs
@@ -698,5 +698,620 @@ pub mod ecs_bundle {
 		pub struct Components {
 			pub settlement: Settlement,
 		}
-	}			
+	}		
+
+	#[system]
+	pub mod move_hero {
+		pub fn execute(ctx: Context<Components>, args: MoveHeroArgs) -> Result<Components> {
+	
+			//todo min cooldown?
+	
+			let hero = &mut ctx.accounts.hero;
+			hero.x = args.x;
+			hero.y = args.y;
+			hero.last_activity = Clock::get()?.unix_timestamp;
+			Ok(ctx.accounts)
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub hero: Hero,
+		}
+	
+		#[arguments]
+		struct MoveHeroArgs {
+			x: i32,
+			y: i32,
+		}
+	}		
+
+	#[system]
+	pub mod repair {
+	    use crate::systems::repair::RepairError;
+
+		pub fn execute(ctx: Context<Components>, args: RepairArgs) -> Result<Components> {
+			let settlement = &mut ctx.accounts.settlement;
+	
+			if args.index as usize >= settlement.buildings.len() {
+				return err!(RepairError::BuildingIndexOutOfRange);
+			}
+	
+			let building = settlement.buildings[args.index as usize];
+			let building_config = &BUILDINGS_CONFIG[building.id as usize];
+	
+			let max_deterioration = config::BASE_DETERIORATION_CAP
+				+ config::DETERIORATION_CAP_RESEARCH_MULTIPLIER
+					* get_research_level(settlement.research, ResearchType::DeteriorationCap);
+	
+			let build_cost = get_construction_cost(
+				settlement.research,
+				building_config.cost_tier,
+				building.level,
+				building.deterioration as f32 / max_deterioration as f32,
+			);
+	
+			if settlement.treasury.wood < build_cost.wood
+				|| settlement.treasury.water < build_cost.water
+				|| settlement.treasury.food < build_cost.food
+				|| settlement.treasury.stone < build_cost.stone
+			{
+				return err!(RepairError::NotEnoughResources);
+			} else {
+				settlement.treasury.wood -= build_cost.wood;
+				settlement.treasury.water -= build_cost.water;
+				settlement.treasury.food -= build_cost.food;
+				settlement.treasury.stone -= build_cost.stone;
+			}
+	
+			//all checks passed
+			settlement.buildings[args.index as usize].deterioration = 0;
+	
+			Ok((ctx.accounts))
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub settlement: Settlement,
+		}
+	
+		#[arguments]
+		struct RepairArgs {
+			index: u8,
+		}
+	}	
+
+	#[system]
+	pub mod research {
+		use crate::{settlement::config::{get_research_level_u8, BITS_PER_RESEARCH, RESEARCH_MASK}, systems::research::ResearchError};
+	
+		pub fn execute(ctx: Context<Components>, args: ResearchArgs) -> Result<Components> {
+			msg!("execute research!: ");
+	
+			let settlement = &mut ctx.accounts.settlement;
+	
+			//would be nice to get the size_of(settlement.research) though
+			if 32 <= (args.research_type * BITS_PER_RESEARCH) as usize {
+				return err!(ResearchError::ResearchIndexOutOfRange);
+			}
+	
+			let mut research_level = get_research_level_u8(settlement.research, args.research_type);
+	
+			let research_cost = config::get_research_cost(args.research_type, research_level);
+	
+			if settlement.treasury.stone < research_cost {
+				return err!(ResearchError::NotEnoughResources);
+			}
+	
+			if research_level >= RESEARCH_MASK {
+				return err!(ResearchError::AlreadyMaxedOut);
+			}
+	
+			settlement.treasury.stone -= research_cost;
+			research_level += 1;
+	
+			let mut research_value = settlement.research;
+	
+			//cut out old value
+			research_value &= !((RESEARCH_MASK as u32) << BITS_PER_RESEARCH * args.research_type);
+	
+			//replace with new one
+			research_value |= (research_level as u32) << BITS_PER_RESEARCH * args.research_type;
+	
+			settlement.research = research_value;
+			// settlement.treasury.gold -= research_cost;
+	
+			if args.research_type == ResearchType::ExtraUnit as u8 {
+				settlement.worker_assignment.push(-1);
+			}
+	
+			Ok(ctx.accounts)
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub settlement: Settlement,
+		}
+	
+		#[arguments]
+		struct ResearchArgs {
+			pub research_type: u8,
+		}
+	}	
+
+	#[system]
+	pub mod reset {
+	    use crate::settlement::config::BuildingType;
+
+		pub fn execute(ctx: Context<Components>, _args: ResetArgs) -> Result<Components> {
+			let clock = Clock::get();
+			let mut now = 0;
+	
+			if clock.is_ok() {
+				now = clock.unwrap().unix_timestamp
+			}
+	
+			ctx.accounts.settlement.buildings = vec![Building {
+				x: 8,
+				y: 8,
+				deterioration: 0,
+				id: BuildingType::TownHall,
+				level: 1,
+				turns_to_build: 0,
+				extraction: get_extraction_cap(1),
+			}];
+	
+			ctx.accounts.settlement.worker_assignment = vec![-1]; //one worker comes as default from town hall
+			ctx.accounts.settlement.treasury = config::INITIAL_TREASURY;
+			ctx.accounts.settlement.environment = config::INITIAL_ENVIRONMENT;
+			ctx.accounts.settlement.time_units = config::INITIAL_TIME_UNITS;
+			ctx.accounts.settlement.faith = config::INITIAL_FAITH;
+			ctx.accounts.settlement.last_time_claim = now;
+			ctx.accounts.settlement.research = 0;
+			ctx.accounts.settlement.quest_claim_status = 0;
+			Ok(ctx.accounts)
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub settlement: Settlement,
+		}
+	
+		#[arguments]
+		struct ResetArgs {}
+	}	
+
+	#[system]
+	pub mod sacrifice {
+    	use crate::systems::sacrifice::SacrificeError;
+	
+		pub fn execute(ctx: Context<Components>, args: SacrificeArgs) -> Result<Components> {
+			let settlement = &mut ctx.accounts.settlement;
+	
+			if settlement.worker_assignment.len() <= args.index as usize {
+				return err!(SacrificeError::LabourIndexOutOfRange);
+			}
+	
+			if settlement.worker_assignment[args.index as usize] < -1 {
+				return err!(SacrificeError::NotRestoredYet);
+			}
+	
+			settlement.worker_assignment[args.index as usize] = config::BASE_DEATH_TIMEOUT
+				+ (config::DEATH_TIMEOUT_RESEARCH_MULTIPLIER
+					* get_research_level(settlement.research, ResearchType::DeathTimeout))
+					as i8;
+			settlement.faith += config::SACRIFICE_FAITH_BOOST;
+			Ok(ctx.accounts)
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub settlement: Settlement,
+		}
+	
+		#[arguments]
+		struct SacrificeArgs {
+			index: u8,
+		}
+	}	
+
+	#[system]
+	pub mod smart_object_init {
+	    use crate::systems::smart_object_init::SmartObjectInitError;
+
+		pub fn execute(ctx: Context<Components>, args: SmartObjectInitArgs) -> Result<Components> {
+			let smart_object_location = &mut ctx.accounts.smart_object_location;
+	
+			//prevent init twice
+			if smart_object_location.entity != Pubkey::default() {
+				return err!(SmartObjectInitError::AlreadyInitialized);
+			}
+	
+			smart_object_location.x = args.x;
+			smart_object_location.y = args.y;
+			smart_object_location.entity = Pubkey::new_from_array(args.entity);
+			Ok(ctx.accounts)
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub smart_object_location: SmartObjectLocation,
+		}
+	
+		#[arguments]
+		struct SmartObjectInitArgs {
+			pub x: i32,
+			pub y: i32,
+	
+			pub entity: [u8; 32],
+		}
+	}
+
+	#[system]
+	pub mod upgrade {
+    	use crate::systems::upgrade::UpgradeError;
+
+		pub fn execute(ctx: Context<Components>, args: UpgradeArgs) -> Result<Components> {
+			let settlement = &mut ctx.accounts.settlement;
+	
+			let index = args.index;
+	
+			if index >= settlement.buildings.len() {
+				return err!(UpgradeError::BuildingIndexOutOfRange);
+			}
+	
+			let building = settlement.buildings[index];
+	
+			if building.turns_to_build > 0 {
+				return err!(UpgradeError::UnderConstruction);
+			}
+	
+			if index > 0 && building.level >= settlement.buildings[0].level {
+				return err!(UpgradeError::TownHallLevelReached);
+			}
+	
+			let building_config = &BUILDINGS_CONFIG[building.id as usize];
+	
+			let build_cost = get_construction_cost(
+				settlement.research,
+				building_config.cost_tier,
+				building.level + 1,
+				1.0,
+			);
+	
+			if settlement.treasury.wood < build_cost.wood
+				|| settlement.treasury.water < build_cost.water
+				|| settlement.treasury.food < build_cost.food
+				|| settlement.treasury.stone < build_cost.stone
+			{
+				return err!(UpgradeError::NotEnoughResources);
+			} else {
+				settlement.treasury.wood -= build_cost.wood;
+				settlement.treasury.water -= build_cost.water;
+				settlement.treasury.food -= build_cost.food;
+				settlement.treasury.stone -= build_cost.stone;
+			}
+	
+			if args.worker_index >= 0 {
+				if settlement.worker_assignment.len() as i16 <= args.worker_index {
+					return err!(UpgradeError::SuppliedWorkerIndexOutOfBounds);
+				}
+				settlement.worker_assignment[args.worker_index as usize] = (index) as i8;
+			}
+	
+			//all checks passed
+			settlement.buildings[index].turns_to_build += get_build_time(
+				settlement.research,
+				building_config.build_time_tier,
+				building.level,
+			);
+	
+			Ok((ctx.accounts))
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub settlement: Settlement,
+		}
+	
+		#[arguments]
+		struct UpgradeArgs {
+			index: usize,
+			worker_index: i16,
+		}
+	}	
+
+	#[system]
+	pub mod wait {
+    use crate::settlement::config::{get_collection_level_multiplier, get_storage_level_multiplier, ENVIRONMENT_MAX};
+
+		pub fn execute(ctx: Context<Components>, args: WaitArgs) -> Result<Components> {
+			let settlement = &mut ctx.accounts.settlement;
+	
+			let time_to_wait = u16::min(args.time, settlement.time_units as u16);
+			settlement.time_units -= time_to_wait as u8;
+	
+			let mut water_storage: u16 = config::INITIAL_TREASURY.water;
+			let mut food_storage: u16 = config::INITIAL_TREASURY.food;
+			let mut wood_storage: u16 = config::INITIAL_TREASURY.wood;
+			let mut stone_storage: u16 = config::INITIAL_TREASURY.stone;
+	
+			//calc current storage capacity for all resources
+			for building in settlement.buildings.to_vec() {
+				if building.turns_to_build > 0 && building.level == 0 {
+					continue;
+				}
+	
+				match building.id {
+					BuildingType::TownHall => {
+						water_storage += config::TOWNHALL_STORAGE_PER_LEVEL * building.level as u16;
+						food_storage += config::TOWNHALL_STORAGE_PER_LEVEL * building.level as u16;
+						wood_storage += config::TOWNHALL_STORAGE_PER_LEVEL * building.level as u16;
+						stone_storage += config::TOWNHALL_STORAGE_PER_LEVEL * building.level as u16;
+					}
+					BuildingType::WaterStorage => {
+						water_storage += config::WATER_STORAGE_PER_LEVEL
+							* get_storage_level_multiplier(building.level);
+					}
+					BuildingType::FoodStorage => {
+						food_storage += config::FOOD_STORAGE_PER_LEVEL
+							* get_storage_level_multiplier(building.level);
+					}
+					BuildingType::WoodStorage => {
+						wood_storage += config::WOOD_STORAGE_PER_LEVEL
+							* get_storage_level_multiplier(building.level);
+					}
+					BuildingType::StoneStorage => {
+						stone_storage += config::STONE_STORAGE_PER_LEVEL
+							* get_storage_level_multiplier(building.level);
+					}
+					_ => {}
+				}
+			}
+	
+			let storage_research =
+				get_research_level(settlement.research, ResearchType::StorageCapacity);
+			let storage_multiplier =
+				1.0 + config::STORAGE_CAPACITY_RESEARCH_MULTIPLIER * storage_research as f32;
+			if storage_multiplier > 1.0 {
+				water_storage = (water_storage as f32 * storage_multiplier).floor() as u16;
+				food_storage = (food_storage as f32 * storage_multiplier).floor() as u16;
+				wood_storage = (wood_storage as f32 * storage_multiplier).floor() as u16;
+				stone_storage = (stone_storage as f32 * storage_multiplier).floor() as u16;
+			}
+	
+			//wells generate water without worker assigned
+			for building in settlement.buildings.to_vec() {
+				if building.turns_to_build > 0 && building.level == 0 {
+					continue;
+				}
+				match building.id {
+					BuildingType::WaterCollector => {
+						let mut collected = 0;
+	
+						if water_storage > settlement.treasury.water {
+							collected = time_to_wait * get_collection_level_multiplier(building.level)
+								+ get_research_level(
+									settlement.research,
+									ResearchType::ResourceCollectionSpeed,
+								) as u16;
+							collected = u16::min(collected, water_storage - settlement.treasury.water);
+						}
+	
+						if collected > 0 {
+							settlement.treasury.water += collected;
+						}
+					}
+					_ => {}
+				}
+			}
+	
+			//process all buildings with allocated worker
+			let mut alive_workers: u16 = 0;
+			for worker_index in 0..settlement.worker_assignment.len() {
+				let building_index = settlement.worker_assignment[worker_index];
+	
+				if building_index >= -1 {
+					alive_workers += 1;
+				}
+	
+				if building_index < 0 {
+					//worker unallocated
+					continue;
+				}
+	
+				let building_index_usize = building_index as usize;
+				let building = settlement.buildings[building_index_usize];
+	
+				let max_deterioration = config::BASE_DETERIORATION_CAP
+					+ config::DETERIORATION_CAP_RESEARCH_MULTIPLIER
+						* get_research_level(settlement.research, ResearchType::DeteriorationCap);
+	
+				if building.deterioration >= max_deterioration {
+					//allocated building broken
+					settlement.worker_assignment[worker_index] = -1;
+					continue;
+				}
+	
+				if building.turns_to_build > 0 {
+					settlement.buildings[building_index_usize].turns_to_build -=
+						u8::min(time_to_wait as u8, building.turns_to_build);
+	
+					if settlement.buildings[building_index_usize].turns_to_build <= 0 {
+						settlement.buildings[building_index_usize].level += 1;
+	
+						settlement.buildings[building_index_usize].extraction +=
+							get_extraction_cap(building.level);
+						if matches!(building.id, BuildingType::TownHall) {
+							settlement.worker_assignment.push(-1);
+						}
+	
+						//finished building anything but food/wood collectors - release workers
+						match settlement.buildings[building_index_usize].id {
+							BuildingType::FoodCollector => {}
+							BuildingType::WoodCollector => {}
+							BuildingType::StoneCollector => {}
+							_ => settlement.worker_assignment[worker_index] = -1,
+						}
+					} else {
+						continue;
+					}
+				}
+	
+				let building_type = building.id;
+	
+				//TODO [CLEANUP] check if this code can be reused (across 3 different resources) (e.g. using array to store resources)
+				match building_type {
+					BuildingType::FoodCollector => {
+						let mut collected = 0;
+	
+						if food_storage > settlement.treasury.food {
+							collected = u16::min(
+								time_to_wait * get_collection_level_multiplier(building.level)
+									+ get_research_level(
+										settlement.research,
+										ResearchType::ResourceCollectionSpeed,
+									) as u16,
+								settlement.environment.food,
+							);
+							collected = u16::min(collected, food_storage - settlement.treasury.food);
+						}
+	
+						if collected > 0 {
+							settlement.environment.food -= collected;
+							settlement.treasury.food += collected;
+						}
+					}
+					BuildingType::WoodCollector => {
+						let mut collected = 0;
+	
+						if wood_storage > settlement.treasury.wood {
+							collected = u16::min(
+								time_to_wait * get_collection_level_multiplier(building.level)
+									+ get_research_level(
+										settlement.research,
+										ResearchType::ResourceCollectionSpeed,
+									) as u16,
+								settlement.environment.wood,
+							);
+							collected = u16::min(collected, wood_storage - settlement.treasury.wood);
+						}
+						if collected > 0 {
+							settlement.environment.wood -= collected;
+							settlement.treasury.wood += collected;
+						}
+					}
+					BuildingType::StoneCollector => {
+						let mut collected = 0;
+	
+						if stone_storage > settlement.treasury.stone {
+							collected = u16::min(
+								time_to_wait * get_collection_level_multiplier(building.level)
+									+ get_research_level(
+										settlement.research,
+										ResearchType::ResourceCollectionSpeed,
+									) as u16,
+								building.extraction,
+							);
+							collected = u16::min(collected, stone_storage - settlement.treasury.stone);
+						}
+						if collected > 0 {
+							settlement.buildings[building_index_usize].extraction -= collected;
+							settlement.treasury.stone += collected;
+						}
+					}
+	
+					_ => {}
+				}
+			}
+	
+			let regeneration_rate = config::get_regeneration_rate(settlement.research);
+	
+			//regeneration sources in environment
+			settlement.environment.food += u16::min(
+				time_to_wait * regeneration_rate,
+				ENVIRONMENT_MAX.food - settlement.environment.food,
+			);
+			settlement.environment.wood += u16::min(
+				time_to_wait * regeneration_rate,
+				ENVIRONMENT_MAX.wood - settlement.environment.wood,
+			);
+	
+			//deteriorate buildings
+			for building in &mut settlement.buildings {
+				if building.turns_to_build == 0 && building.deterioration < u8::MAX {
+					building.deterioration += time_to_wait as u8;
+				}
+			}
+	
+			if settlement.treasury.water < alive_workers || settlement.treasury.food < alive_workers {
+				//kill one
+				for i in 0..settlement.worker_assignment.len() {
+					if (settlement.worker_assignment[i]) >= -1 {
+						settlement.worker_assignment[i] = config::BASE_DEATH_TIMEOUT
+							+ (config::DEATH_TIMEOUT_RESEARCH_MULTIPLIER
+								* get_research_level(settlement.research, ResearchType::DeathTimeout))
+								as i8;
+	
+						alive_workers -= 1;
+						break;
+					}
+				}
+			}
+	
+			let consumption_rate: u16 = (alive_workers
+				- u16::min(
+					alive_workers,
+					get_research_level(settlement.research, ResearchType::Consumption) as u16,
+				))
+				* time_to_wait;
+	
+			settlement.treasury.water -= u16::min(settlement.treasury.water, consumption_rate);
+			settlement.treasury.food -= u16::min(settlement.treasury.food, consumption_rate);
+	
+			let mut i: usize = 0;
+	
+			//restore sacrificed worker
+			for building_index in settlement.worker_assignment.to_vec() {
+				if building_index < -1 {
+					settlement.worker_assignment[i] += time_to_wait as i8;
+					if settlement.worker_assignment[i] > -1 {
+						settlement.worker_assignment[i] = -1;
+					}
+				}
+	
+				i += 1;
+			}
+	
+			//calc faith as a lerp to 'runway'
+			let mut runway = 0;
+			if alive_workers > 0 {
+				runway = u16::min(settlement.treasury.food, settlement.treasury.water)
+					/ alive_workers as u16;
+			}
+	
+			msg!("runway {}", runway);
+	
+			if settlement.faith >= config::FAITH_TO_RUNWAY_LERP_PER_TURN
+				&& runway < settlement.faith as u16
+			{
+				settlement.faith -= config::FAITH_TO_RUNWAY_LERP_PER_TURN;
+			} else if settlement.faith < u8::MAX - config::FAITH_TO_RUNWAY_LERP_PER_TURN
+				&& runway > settlement.faith as u16
+			{
+				settlement.faith += config::FAITH_TO_RUNWAY_LERP_PER_TURN;
+			}
+	
+			Ok(ctx.accounts)
+		}
+	
+		#[system_input]
+		pub struct Components {
+			pub settlement: Settlement,
+		}
+	
+		#[arguments]
+		struct WaitArgs {
+			time: u16,
+		}
+	}	
 }
