--- programs-ecs/systems/wait/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "wait"
-version = "0.1.10"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "wait"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build", "ecs-bundle/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-anchor-spl.workspace = true
-serde.workspace = true
-ecs-bundle.workspace = true
